fix themeAttrDetect and add trackChangesHTMLElement
