+++ package/src/scripts/utils/observeHtmlElement.ts
@@ -1,0 +1,15 @@
+const trackChangesHTMLElement = (htmlEl: HTMLElement, attr: string, actions: () => void) => {
+  const changes = (mutationsList: MutationRecord[]) => {
+    for (let i = 0; i < mutationsList.length; i++) {
+      const record = mutationsList[i];
+      if (record.attributeName === attr) {
+        actions();
+        break;
+      }
+    }
+  };
+  const observer = new MutationObserver(changes);
+  observer.observe(htmlEl, { attributes: true });
+};
+
+export default trackChangesHTMLElement;
